fixed error on green geeks server
originally i was reusing the script that sends the action data to the
server to send the checkbox data,  and call a second script to save the
email address. i combined these into one so that we are only making one
call to the server. the green geeks server would occasionally error out
when make these two calls.